refactor: clean up logs and functions (#4)
