+++ das/eigenda/eigenda.go
@@ -3,14 +3,18 @@ package eigenda
 import (
 	"bytes"
 	"context"
+	"crypto/sha256"
 	"crypto/tls"
 	"encoding/binary"
+	"encoding/hex"
 	"errors"
 	"fmt"
 	"time"
 
 	"github.com/Layr-Labs/eigenda/api/grpc/disperser"
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/log"
+	"github.com/offchainlabs/nitro/arbutil"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 )
@@ -77,7 +81,6 @@ func (b *EigenDARef) Deserialize(data []byte) error {
 type EigenDA struct {
 	client disperser.DisperserClient
 }
-
 
 func NewEigenDA(rpc string) (*EigenDA, error) {
 	creds := credentials.NewTLS(&tls.Config{
@@ -171,4 +174,41 @@ func (e *EigenDA) Serialize(eigenDARef *EigenDARef) ([]byte, error) {
 	}
 	serializedBlobPointerData := buf.Bytes()
 	return serializedBlobPointerData, nil
+}
+
+func RecoverPayloadFromEigenDABatch(ctx context.Context,
+	batchNum uint64,
+	sequencerMsg []byte,
+	daReader EigenDAReader,
+	preimages map[arbutil.PreimageType]map[common.Hash][]byte,
+) ([]byte, error) {
+	log.Info("Start recovering payload from eigenda: ", "data", hex.EncodeToString(sequencerMsg))
+	var shaPreimages map[common.Hash][]byte
+	if preimages != nil {
+		if preimages[arbutil.Sha2_256PreimageType] == nil {
+			preimages[arbutil.Sha2_256PreimageType] = make(map[common.Hash][]byte)
+		}
+		shaPreimages = preimages[arbutil.Sha2_256PreimageType]
+	}
+	// 00000020
+	// 91c127a758d669ce7c8ed915679653e87bf1dfbcf54d028c522d129c482c897d
+	var daRef EigenDARef
+	daRef.BlobIndex = binary.BigEndian.Uint32(sequencerMsg[:4])
+	daRef.BatchHeaderHash = sequencerMsg[4:]
+	log.Info("Data pointer: ", "info", hex.EncodeToString(daRef.BatchHeaderHash), "index", daRef.BlobIndex)
+	data, err := daReader.QueryBlob(ctx, &daRef)
+	if err != nil {
+		log.Error("Failed to query data from EigenDA", "err", err)
+		return nil, err
+	}
+	// log.Info("data: ", "info", hex.EncodeToString(data))
+	// record preimage data
+	log.Info("Recording preimage data for EigenDA")
+	shaDataHash := sha256.New()
+	shaDataHash.Write(sequencerMsg)
+	dataHash := shaDataHash.Sum([]byte{})
+	if shaPreimages != nil {
+		shaPreimages[common.BytesToHash(dataHash)] = data
+	}
+	return data, nil
 }
